--- tests/unit/query_graph/adapter/test_adapter_common.py
@@ -1,41 +1,0 @@
-from typing import cast
-
-import pytest
-from sqlglot import parse_one
-from sqlglot.expressions import Select
-
-from featurebyte.enum import SourceType
-from featurebyte.query_graph.node.schema import TableDetails
-from featurebyte.query_graph.sql.adapter import get_sql_adapter
-
-
-@pytest.mark.parametrize(
-    "source_type, expected",
-    [
-        (
-            SourceType.SNOWFLAKE,
-            'CREATE TABLE "db1"."schema1"."table1" AS SELECT * FROM A',
-        ),
-        (
-            SourceType.SPARK,
-            "CREATE TABLE `db1`.`schema1`.`table1` USING DELTA TBLPROPERTIES ('delta.columnMapping.mode'='name', 'delta.minReaderVersion'='2', 'delta.minWriterVersion'='5') AS SELECT * FROM A",
-        ),
-        (
-            SourceType.DATABRICKS,
-            "CREATE TABLE `db1`.`schema1`.`table1` USING DELTA TBLPROPERTIES ('delta.columnMapping.mode'='name', 'delta.minReaderVersion'='2', 'delta.minWriterVersion'='5') AS SELECT * FROM A",
-        ),
-    ],
-)
-def test_create_table_as(source_type, expected):
-    """
-    Test create_table_as for Adapter
-    """
-
-    table_details = TableDetails(
-        database_name="db1",
-        schema_name="schema1",
-        table_name="table1",
-    )
-    expr = parse_one("SELECT * FROM A")
-    new_expr = get_sql_adapter(source_type).create_table_as(table_details, cast(Select, expr))
-    assert new_expr.sql(dialect=source_type).strip() == expected
